+++ src/components/app/app.module.css
@@ -1,3 +1,3 @@
-.App {
+.app {
   text-align: center;
 }
